fix from Tim Oess (Ulm) to timestamping with timestamp multiplier to avoid only even values of timestamps. M.Sc. Timo Oess Appl. Cognitive Psychology Faculty for Computer Science, Engineering, and Psychology Ulm University Albert-Einstein-Allee 43 89069 Ulm

+++ src/net/sf/jaer/eventio/AEUnicastOutput.java
@@ -212,12 +212,12 @@ public class AEUnicastOutput implements AEUnicastSettings {
                         if (use4ByteAddrTs) {
                             currentBuf.putInt(addr[i]);
                             if (timestampsEnabled) {
-                                currentBuf.putInt((int) (timestampMultiplierReciprocal * t));
+                                currentBuf.putInt((int) (timestampMultiplierReciprocal * (double)t));
                             }
                         } else {
                             currentBuf.putShort((short) addr[i]);
                             if (timestampsEnabled) {
-                                currentBuf.putInt((int) (timestampMultiplierReciprocal * t));
+                                currentBuf.putInt((int) (timestampMultiplierReciprocal * (double)t));
                             }
                         }
                         //No payload.
@@ -243,22 +243,22 @@ public class AEUnicastOutput implements AEUnicastSettings {
                         if (use4ByteAddrTs) {
                             currentBuf.putInt(swab(addr[i]));
                             if (timestampsEnabled) {
-                                currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * t)));
+                                currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * (double)t)));
                             }
                         } else {
                             currentBuf.putShort((short) swab(addr[i]));
                             if (timestampsEnabled) {
-                                currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * t)));
+                                currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * (double)t)));
                             }
                         }
                     } else if (use4ByteAddrTs) {
                         if (timestampsEnabled) {
-                            currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * t)));
+                            currentBuf.putInt(swab((int) (timestampMultiplierReciprocal * (double)t)));
                         }
                         currentBuf.putInt(swab(addr[i]));
                     } else {
                         if (timestampsEnabled) {
-                            currentBuf.putShort((short) swab((int) (timestampMultiplierReciprocal * t)));
+                            currentBuf.putShort((short) swab((int) (timestampMultiplierReciprocal * (double)t)));
                         }
                         currentBuf.putInt(swab(addr[i]));
                     }
